+++ src/components/Control/Control.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import {Icon} from '@yandex-cloud/uikit';
-import {SVGIconData} from '@yandex-cloud/uikit/build/esm/components/Icon/types';
+import {Icon} from '@gravity-ui/uikit';
+import {SVGIconData} from '@gravity-ui/uikit/build/esm/components/Icon/types';
 
 import {block} from '../../utils';