+++ src/components/ToggleArrow/ToggleArrow.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Icon} from '@yandex-cloud/uikit';
+import {Icon} from '@gravity-ui/uikit';
 
 import {block} from '../../utils';
 import {Chevron, NavigationChevron} from '../../icons';